requirements/local.txt
-r base.txt
coverage==3.7.1
django-debug-toolbar==1.3.0
flake8==2.4.0
pylint==1.4.3
selenium==2.45.0
sphinx==1.3.1
